:sparkles: add: nft/list apit
