--- server/src/controller/nft.js
@@ -1,11 +1,0 @@
-const mint = async (req, res, next) => {};
-
-const nftExchange = async (req, res, next) => {};
-
-const nftBuy = async (req, res, next) => {};
-
-const nftList = async (req, res, next) => {
-  const { user_id } = req.body;
-};
-
-const nftSell = async (req, res, next) => {};
