+++ server/src/db_Process/nft.db.js
@@ -1,0 +1,7 @@
+import Market_nft from "../../models/market_nft";
+
+//get-nft/list
+exports.marketNft_list = async () => {
+  const result = await Market_nft.findAll();
+  return result;
+};
